+++ src/ctests/sdsc2.c
@@ -37,7 +37,7 @@ int main(int argc, char **argv)
    double spread[MAXEVENTS];
    int nevents = MAXEVENTS;
    int eventset = PAPI_NULL;
-   int events[MAXEVENTS];
+   long long events[MAXEVENTS];
 
    events[0] = PAPI_FP_INS;
    events[1] = PAPI_TOT_INS;
